mito-ai: refactor prompt_builders
